Rename SimpleInputMechanism to StringInputMechanism and PositionBufferedInputMechanism to PositionBufferedStringInputMechanism
